+++ .gitignore
@@ -29,5 +29,6 @@ t/fill*
 # example output
 examples/**/out
 *.swp
+built
