+++ index.js
@@ -1,2 +1,2 @@
 'use strict'
-module.exports = require('./src')
+module.exports = require('./built')
