+++ jest.config.js
@@ -2,6 +2,10 @@ module.exports = {
 //  preset: "ts-jest",
   setupFilesAfterEnv: ["jest-extended"],
   testEnvironment: "node",
+  "testPathIgnorePatterns": ["/src/"],
+  "modulePathIgnorePatterns": [
+    "<rootDir>/src"
+  ]
 //  transform: {
 //    "^.+\\.tsx?$": "ts-jest",
 //  },
