+++ package.json
@@ -5,7 +5,12 @@
   "main": "index.js",
   "scripts": {
     "postversion": "git commit package.json CHANGELOG.md -m \"v$npm_package_version\" && npm run tag && git push && git push --tags",
-    "test": "jest"
+    "test": "jest",
+    "watch": "run-p -r watch:*",
+    "watch:ts": "tsc -w --pretty",
+    "watch:tests": "jest --watch -t \"test\"",
+    "watch:pegsj": "nodemon"
+
   },
   "bin": {
     "pod6html": "bin/pod6html"
@@ -32,6 +37,7 @@
     "jest": "^24.9.0",
     "jest-extended": "^0.11.2",
     "nodemon": "^2.0.2",
+    "npm-run-all": "^4.1.5",
     "pegjs": "^0.10.0"
   }
 }
